refactor(gui): update paths after moving utility files
