+++ src/gui/components/config/GraphComponent.h
@@ -8,9 +8,9 @@
 #pragma once
 
 #include "../../../Interpolator.h"
+#include "../../../utility/PointComparator.h"
+#include "../../../utility/clip.h"
 #include "../../appearance/Colours.h"
-#include "../../utility/PointComparator.h"
-#include "../../utility/clip.h"
 #include <JuceHeader.h>
 #include <memory>
